Preps Forge for Docker Integration (#3)
* Created standalone Dockerfile for using Alpine and Forge interactively.

* Added POSIX grep -r vs. -R fix.

* Added POSIX cmp -s vs. --silent.

* Added temp branch targeting for Dockerfile branch.

* Updated the Dockerfile bash example with ssh mounting.

* Replaced cache key path generation with a checksum value.

Forge version_control now caches files with their filename as the checksum. This enables O(1) lookup of all files, eliminates collisions, and simplifies the logic. This will also open up the possibility of mounting the cache for containerization while preserving the identity of the cache.

* Created Dockerfiles directory with executable and interactive versions of the forge Dockerfiles.

* Updated interactive Dockerfile to install forge globally.

* Updated bash forge alias doc for both Dockerfiles.

* Added bust_stale_cache function for busting the forge cache after 30 days.

* Updated README to include new cache busting info.

* Updated file caching to merely touch a file with the checksum as its name.

This will save disk space, as the only info required by the caching is the checksum: the file contents are meaningless.

* Removed Dockerfiles and added README reference to docker-forge.